app/models/model_manager.py
"""
Simple model manager for LLM selection and fallback handling.
"""
import os
import yaml
import json
import re
from typing import Optional, Dict, Any
from dataclasses import dataclass

@dataclass
class ModelConfig:
    """Model configuration data."""
    name: str
    provider: str

class ModelManager:
    """
    Manages LLM model selection and fallback behavior.
    
    Provides a simple interface to:
    1. Access the current model configuration
    2. Handle fallback to backup model on errors
    3. Reset to primary model when desired
    """
    
    def __init__(self, config_path: str):
        """
        Initialize the model manager.
        
        Args:
            config_path: Path to YAML configuration file
            
        Raises:
            ValueError: If configuration is invalid or API keys are missing
        """
        self._load_config(config_path)
        self._validate_api_keys()
        self._using_primary = True

    def _load_config(self, config_path: str) -> None:
        """Load model configuration from YAML file."""
        with open(config_path, 'r') as f:
            config = yaml.safe_load(f)

        if not config or 'models' not in config:
            raise ValueError("Invalid configuration file")

        models = config['models']
        if 'primary' not in models or 'backup' not in models:
            raise ValueError("Configuration must specify primary and backup models")

        self.primary = ModelConfig(**models['primary'])
        self.backup = ModelConfig(**models['backup'])

    def _validate_api_keys(self) -> None:
        """Validate required API keys are present."""
        required_keys = {
            'openai': 'OPENAI_API_KEY',
            'anthropic': 'ANTHROPIC_API_KEY'
        }

        # Check for primary model's API key
        key_name = required_keys.get(self.primary.provider)
        if not key_name or not os.getenv(key_name):
            raise ValueError(f"Missing API key for {self.primary.provider}")

        # Check for backup model's API key
        key_name = required_keys.get(self.backup.provider)
        if not key_name or not os.getenv(key_name):
            raise ValueError(f"Missing API key for {self.backup.provider}")

    @property
    def current_model(self) -> str:
        """Get the current model name."""
        return self.primary.name if self._using_primary else self.backup.name

    @property
    def current_provider(self) -> str:
        """Get the current provider name."""
        return self.primary.provider if self._using_primary else self.backup.provider

    @property
    def api_key(self) -> str:
        """Get the API key for the current provider."""
        key_mapping = {
            'openai': 'OPENAI_API_KEY',
            'anthropic': 'ANTHROPIC_API_KEY'
        }
        provider = self.current_provider
        key_name = key_mapping.get(provider)
        if not key_name:
            raise ValueError(f"Unsupported provider: {provider}")
        
        key = os.getenv(key_name)
        if not key:
            raise ValueError(f"Missing API key for {provider}")
        return key

    def handle_model_error(self) -> None:
        """Switch to backup model after an error."""
        self._using_primary = False

    def reset_to_primary(self) -> None:
        """Reset to using the primary model."""
        self._using_primary = True
        
    def generate_review(self, prompt: str) -> str:
        """
        Generate a code review using the current model.
        
        Args:
            prompt: The formatted prompt to send to the model
            
        Returns:
            str: JSON string containing the review results
            
        Raises:
            ValueError: If model generation fails
        """
        # Determine review type from prompt
        if "MERGED REVIEW TO ANALYZE:" in prompt:
            # Return mock final review
            mock_review = {
                "final_assessment": {
                    "final_score": 8.3,
                    "quality_breakdown": {
                        "code_quality": 8.5,
                        "architecture": 7.9,
                        "maintainability": 8.2,
                        "scalability": 8.0
                    },
                    "recommendations": [
                        "Implement comprehensive architectural documentation with business impact analysis",
                        "Establish cross-team code review practices to maintain quality",
                        "Develop automated testing strategy with clear ROI metrics"
                    ],
                    "risk_analysis": {
                        "high_priority": [
                            "Architectural documentation gaps",
                            "Cross-component integration points"
                        ],
                        "medium_priority": [
                            "Code duplication across services",
                            "Inconsistent error handling"
                        ],
                        "low_priority": [
                            "Minor style inconsistencies",
                            "Documentation updates needed"
                        ]
                    }
                },
                "action_plan": {
                    "immediate_actions": [
                        "Document critical architectural decisions",
                        "Implement integration test suite"
                    ],
                    "short_term_goals": [
                        "Refactor shared utilities",
                        "Standardize error handling"
                    ],
                    "long_term_improvements": [
                        "Migrate to microservices architecture",
                        "Implement continuous architecture reviews"
                    ]
                },
                "summary": "The codebase demonstrates strong fundamentals with opportunities for architectural improvements"
            }
        elif "BATCH REVIEWS TO MERGE:" in prompt:
            # Return mock merged batch review
            mock_review = {
                "merged_analysis": {
                    "overall_quality_score": 8.2,
                    "architectural_alignment_score": 7.9,
                    "integration_impact_score": 8.5,
                    "key_findings": {
                        "strengths": [
                            "Consistent architectural patterns",
                            "Good separation of concerns",
                            "Strong type safety practices"
                        ],
                        "concerns": [
                            "Mixed language integration points",
                            "Some architectural inconsistencies"
                        ],
                        "risks": [
                            "Cross-language type safety",
                            "Architectural drift between components"
                        ]
                    }
                },
                "recommendations": {
                    "architectural_improvements": [
                        "Refactor towards layered architecture pattern",
                        "Implement architectural decision records",
                        "Strengthen architectural boundaries"
                    ],
                    "integration_considerations": [
                        "Implement strict type checking",
                        "Document cross-component interfaces"
                    ],
                    "priority_actions": [
                        "Define clear architectural guidelines",
                        "Add integration tests"
                    ]
                },
                "summary": "Overall strong codebase with good architectural foundation, some opportunities for improvement"
            }
        elif "FILES TO REVIEW:" in prompt:
            # Extract file paths for batch review
            file_paths = re.findall(r'File: (.+?)\n', prompt)
            
            # Return mock batch review
            mock_review = {
                "batch_analysis": {
                    "files_reviewed": file_paths,
                    "consistency_score": 8.5,
                    "pattern_quality": 7.8,
                    "cohesion_rating": 8.2,
                    "findings": {
                        "patterns_identified": [
                            "Consistent function naming",
                            "Mixed language usage patterns"
                        ],
                        "consistency_issues": [
                            "Minor style variations"
                        ],
                        "cohesion_concerns": [
                            "Some duplicate utilities"
                        ]
                    }
                },
                "recommendations": {
                    "pattern_improvements": [
                        "Extract common utilities",
                        "Standardize error handling"
                    ],
                    "consistency_fixes": [
                        "Apply consistent naming",
                        "Standardize file structure"
                    ],
                    "cohesion_enhancements": [
                        "Create shared module",
                        "Improve interfaces"
                    ]
                }
            }
        else:
            # Extract file path for individual review
            path_match = re.search(r'Path: (.+)\n', prompt)
            if not path_match:
                raise ValueError("Invalid prompt format: missing file path")
            file_path = path_match.group(1)
            
            # Return mock individual file review
            mock_review = {
                "file_scores": {
                    file_path: {
                        "readability": 8.5,
                        "maintainability": 7.8,
                        "complexity": 6.5,
                        "coding_standards": 9.0,
                        "documentation": 7.0,
                        "security": 8.0,
                        "performance": 8.5,
                        "reusability": 7.5,
                        "error_handling": 8.0,
                        "test_coverage": 6.0,
                        "notes": "Well-structured code with good practices."
                    }
                },
                "overall_review": {
                    "total_score": 7.7,
                    "strengths": ["Clean code", "Good organization"],
                    "concerns": ["Could use more comments"],
                    "hiring_confidence": 8.0,
                    "risks": ["Minor documentation gaps"],
                    "summary": "Overall solid code quality with room for minor improvements."
                }
            }
            
        return json.dumps(mock_review)
